RELATORIO FINALIZADO

+++ src/main/java/br/com/uniamerica/estacionamento/entity/Condutor.java
@@ -25,7 +25,7 @@ public class Condutor extends AbstractEntity{
 
     @Getter@Setter
     @Column(name = "tempo_gasto")
-    private Integer tempoPago;
+    private BigDecimal tempoPago;
 
     @Getter@Setter
     @Column(name = "tempo_desconto")
